+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## Bug Fixes:
 
+- [#364](http://github.com/Nelson-numerical-software/nelson/issues/364): `isinf`, `isnan`, `conj`, `double`, `single`, `real`, `imag` are faster.
+
 - [#361](http://github.com/Nelson-numerical-software/nelson/issues/361): `abs` is faster.
 
 - [#360](http://github.com/Nelson-numerical-software/nelson/issues/360): `ctranpose` and `transpose` are faster.